add support for configuration
